Correct some premature state transition optimizations and hidden state stuff in action menu
