+++ Assets/Scripts/Contexts/Battle/Commands/CombatActionSelectedCommand.cs
@@ -1,7 +1,5 @@
 ﻿using Contexts.Battle.Models;
-using Contexts.Battle.Signals;
 using Models.Combat;
-using Models.Fighting;
 using Models.Fighting.Skills;
 using strange.extensions.command.impl;
 
@@ -13,9 +11,6 @@ namespace Contexts.Battle.Commands {
         [Inject]
         public BattleViewState Model { get; set; }
 
-        [Inject]
-        public NewMoveRangeSignal NewMoveRangeSignal { get; set; }
-
         public override void Execute() {
             if (ActionType == CombatActionType.Attack) {
                 // TODO: Check if we should use melee or ranged.
@@ -25,13 +20,6 @@ namespace Contexts.Battle.Commands {
                 // TODO: Determine special skill type
             } else if (ActionType == CombatActionType.Move) {
                 Model.State = BattleUIState.SelectingMoveLocation;
-
-                var map = Model.Map;
-                var origin = Model.SelectedCombatant.Position;
-                var moveRange = Model.SelectedCombatant.GetAttribute(Attribute.AttributeType.Move);
-                var squares = map.BreadthFirstSearch(origin, moveRange.Value, false);
-
-                NewMoveRangeSignal.Dispatch(squares);
             } else if (ActionType == CombatActionType.Fight) { 
                 Model.State = BattleUIState.SelectingFightAction;
             } else {
